Moved test files to appropriate packages and amended names
