+++ src/test/java/io/samsquamptch/afterclass/service/LessonServiceTests.java
@@ -1,0 +1,4 @@
+package io.samsquamptch.afterclass.service;
+
+public class LessonServiceTests {
+}
