+++ src/test/java/io/samsquamptch/afterclass/service/UserServiceTests.java
@@ -1,0 +1,4 @@
+package io.samsquamptch.afterclass.service;
+
+public class UserServiceTests {
+}
